Add androidTarget and Locale for android
