+++ buildSrc/build.gradle.kts
@@ -7,9 +7,11 @@ repositories {
     mavenCentral()
     // so that external plugins can be resolved in dependencies section
     gradlePluginPortal()
+    google()
 }
 
 dependencies {
    implementation(libs.kotlinGradlePlugin)
    implementation(libs.dokkaPlugin)
+   implementation(libs.androidGradlePlugin)
 }
