+++ i18n4k-core/src/jvmAndroidMain/kotlin/de/comahe/i18n4k/Locale.jvmAndroid.kt
@@ -1,0 +1,36 @@
+package de.comahe.i18n4k
+
+import de.comahe.i18n4k.strings.capitalize
+
+/** Class representing a locale */
+actual typealias Locale = java.util.Locale
+
+actual fun createLocale(
+    language: String,
+    script: String?,
+    country: String?,
+    variant: String?,
+    extensions: Map<Char, String>?
+): Locale {
+    val builder = java.util.Locale.Builder()
+
+    builder.setLanguage(language.lowercase())
+    if (script != null)
+        builder.setScript(script.capitalize())
+    if (country != null)
+        builder.setRegion(country.uppercase())
+    if (variant != null)
+        builder.setVariant(variant.lowercase().take(8))
+    extensions?.forEach { (key, value) ->
+        builder.setExtension(key.lowercaseChar(), value.lowercase())
+    }
+    return builder.build()
+}
+
+@Suppress("TYPEALIAS_EXPANSION_DEPRECATION")
+val jvmSystemLocale: Locale
+    get() {
+        val userLanguage = System.getProperty("user.language")
+            ?: return Locale.getDefault()
+        return Locale(userLanguage)
+    }
