+++ libs.versions.toml
@@ -18,6 +18,8 @@ slf4j = "1.7.36"
 logback = "1.2.11"
 # https://junit.org/junit4/
 junit = "4.13.2"
+#https://developer.android.com/jetpack/androidx/releases/core
+androidXcore = "1.15.0"
 
 ###############  Plugins ###################
 
@@ -47,6 +49,7 @@ kotlinPoet = { module = "com.squareup:kotlinpoet", version.ref = "kotlinPoet" }
 slf4jApi = { module = "org.slf4j:slf4j-api", version.ref = "slf4j" }
 logbackClassic = { module = "ch.qos.logback:logback-classic", version.ref = "logback" }
 junit = { module = "junit:junit", version.ref = "junit" }
+androidxCoreKtx = { module = "androidx.core:core-ktx", version.ref = "androidXcore" }
 
 
 [plugins]
